backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
pydantic[email]
email-validator
python-multipart
pandas
openpyxl
pdfplumber
python-docx
python-dotenv
python-telegram-bot
httpx
langchain-core
langchain-groq
langgraph
